Updated the project structure
The new structure is:
application
frontends
  api
  slack-integration
search-providers
  api
  google-image-search
service-layer
  api
  service

resolves #11
