--- api/src/main/java/com/github/tsar_industries/slashgif/api/image_search/Image.java
@@ -1,5 +1,0 @@
-package com.github.tsar_industries.slashgif.api.image_search;
-
-public class Image
-{
-}
