+++ service-layer/api/src/main/java/com/github/tsar_industries/slashgif/service_layer/api/Image.java
@@ -1,0 +1,5 @@
+package com.github.tsar_industries.slashgif.service_layer.api;
+
+public class Image
+{
+}
