+++ service-layer/api/src/main/java/com/github/tsar_industries/slashgif/service_layer/api/Service.java
@@ -1,0 +1,9 @@
+package com.github.tsar_industries.slashgif.service_layer.api;
+
+
+import java.util.Optional;
+
+public interface Service
+{
+	Optional<Image> getImage(String query);
+}
