--- web/src/main/java/com/github/tsar_industries/slashgif/web/SlashgifConfiguration.java
@@ -1,10 +1,0 @@
-package com.github.tsar_industries.slashgif.web;
-
-import io.dropwizard.Configuration;
-
-/**
- * This is the java-representation of the yaml-config file
- */
-public class SlashgifConfiguration extends Configuration
-{
-}
